+++ packages/definitions/src/typings/studio/index.ts
@@ -1,5 +1,4 @@
 export * from './asset';
 export * from './release';
 export * from './resource';
-export * from './terminal';
 export * from './workspace';