+++ packages/ap-core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @recative/ap-core
 
+## 0.2.0
+
+### Minor Changes
+
+- BREAKING: Change preload strategy
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @recative/definitions@0.4.0
+  - @recative/act-protocol@0.2.7
+  - @recative/resource-bridge@0.2.3
+
 ## 0.1.0
 
 ### Minor Changes