+++ packages/core-manager/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.4.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @recative/definitions@0.4.0
+  - @recative/act-protocol@0.2.7
+  - @recative/resource-loader@0.2.2
+
+## 0.4.4
+
+### Patch Changes
+
+- chore: Update configuration of slowTaskQueue
+
 ## 0.4.3
 
 ### Patch Changes